+++ src/Mzinga/AppInfo.cs
@@ -12,7 +12,7 @@ using System.Runtime.InteropServices;
 
 namespace Mzinga
 {
-    public class AppInfo
+    public static class AppInfo
     {
         public static Assembly Assembly => _assembly ??= Assembly.GetEntryAssembly() ?? Assembly.GetExecutingAssembly();
         private static Assembly? _assembly = null;
@@ -79,5 +79,10 @@ namespace Mzinga
             @"The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.",
             @"THE SOFTWARE IS PROVIDED ""AS IS"", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE."
         };
+
+        public static string FormattedLicensesText => string.Join(Environment.NewLine + Environment.NewLine,
+            "# Licenses #",
+            $"## {HiveProduct} ##", HiveCopyright, HiveLicense,
+            $"## {Product} ##", MitLicenseName, Copyright, MitLicenseBody);
     }
 }